requirements.txt
dotenv
google-generativeai
mypy>=1.18.1
pylint>=3.3.8
trafilatura